[GENERATED] Initial Commit
This commit is the output of the following command (`rails@7.2.0.alpha`):

```
rails new railsconf-2024 \
  --main \
  --database=sqlite3 \
  --asset-pipeline=propshaft \
  --javascript=esbuild \
  --css=tailwind \
  --skip-jbuilder \
  --skip-action-mailbox \
  --skip-spring
```

This command skips a few Rails-provided dependencies that the project
won't be utilizing.

It sets [SQLite][] as the application's database, configures the
project's JavaScript to be built with [esbuild][] through [Propshaft][],
the project's CSS to be built with [Tailwind CSS][],
and makes [StimulusJS][] available on the project's client-side.

[SQLite]: https://www.sqlite.org
[esbuild]: https://esbuild.github.io
[Propshaft]: https://github.com/rails/propshaft
[Tailwind CSS]: https://tailwindcss.com
[StimulusJS]: https://stimulus.hotwired.dev