Prepare the application for load benchmarking
In order to assess our progress, we need some way to benchmark our application. The README details how to use the `oha` load testing CLI utility along with the new `/benchmarking` routes to assess some key metrics for our application. A baseline `production.sqlite3` database file is also provided so that `git clone` is truly all that is needed to get started.