Prepare the application for load benchmarking
In order to assess our progress, we need some way to benchmark our application. The README details how to use the `oha` load testing CLI utility along with the new `/benchmarking` routes to assess some key metrics for our application. A baseline `production.sqlite3` database file is also provided so that `git clone` is truly all that is needed to get started.

Run the baseline load tests

In order to guide the first phase of our work, we need a clear picture of where the pain points are in the out-of-the-box experience of running SQLite on Rails. Running the `posts_index` and `post_create` benchmarking endpoints through our load testing tool reveals 4 immediate pain points:

1. There are *way too many* errored responses. Looking at the logs, you will see that all of these errors are `SQLite3::BusyException: database is locked` exceptions.
2. The slowest request took *50 times* longer than the average request. This is a sign that the app is not handling the load well at all.
3. Even on our high-powered laptop over localhost, our server can only support ~40 requests per second on the `posts_index` endpoint. This is a low number, and should be higher.
4. The p99 response time is over 5 seconds. This is a very high number, and it is likely that users will not be happy with the performance of the app.