Control compilation flags for SQLite
The `sqlite3-ruby` gem allows us to control the compilation flags used to compile the SQLite executable. We can set the compilation flags via the `BUNDLE_BUILD__SQLITE3` environment variable set in the `.bundle/config` file. Typically, this file is removed from source control, but we add it back to make this app more portable. Note, however, that this does restrict individual configuration of Bundler.

In order to ensure that SQLite is compiled from source, we need to specify in the `Gemfile` that the SQLite gem should use the `ruby` platform version.