Setup Litestream for point-in-time-backups
In order to ensure that our SQLite on Rails application runs smoothly and resiliently, we need to ensure that we have a backup strategy for our SQLite database(s). Among the various options, I prefer the [Litestream utility][], which continuously streams SQLite changes to a remote S3-compatible storage provider.

For this workshop, we will run a local S3-compatible storage service called [MinIO][] and connect our Litestream instance with that local server.

With everything setup, we can run the Litestream replication process and observe our `storage/production.sqlite3` being replicated to the configured MinIO bucket.

[Litestream utility]: https://litestream.io
[MinIO]: https://github.com/minio/minio