Verify that Litestream backups are restorable and fresh
Setting up a backup strategy is important, but without confirmation that your backups are restorable and fresh it is relatively useless. So, after setting up Litestream and a local MinIO instance, we need to determine how to verify that our Litestream replication process is working. Ideally, we also want to setup an automated process that will regularly check that our backups are both fresh and restorable.