Install and use SQLite extensions
One of the best way to enhance your experience with SQLite is to install and use any of the many extensions in the SQLite ecosystem. The `sqlpkg` utility is a package manager for SQLite extensions. The Ruby gem that provides the cross-platform precompiled binaries for the utility, alongside the enhanced adapter, allow the application to install, load, and use any of the extensions available through the package manager. In order to ensure that we can download and use the extensions across separate development and production operating systems, we need to setup the application to store which extensions are needed for the application, which configurations should load them, but not store the executables themselves in the repository.