Control SQLite compilation
Because SQLite is simply a single executable, it is easy to control the actual compilation of SQLite. The `sqlite3-ruby` gem allows us to control the compilation flags used to compile the SQLite executable. Bundler allows you set set such configuration directly. SQLite recommends a set of 12 flags, but some omit features that the `sqlite3-ruby` gem needs and some are features that are useful for web applications. We recommend a set of 7:

-DSQLITE_DQS=0
-DSQLITE_THREADSAFE=0
-DSQLITE_DEFAULT_MEMSTATUS=0
-DSQLITE_LIKE_DOESNT_MATCH_BLOBS
-DSQLITE_MAX_EXPR_DEPTH=0
-DSQLITE_OMIT_SHARED_CACHE
-DSQLITE_USE_ALLOCA