Working with branch-specific databases
Because SQLite stores your entire database in literal files on disk and only runs embedded in your application process, databases are very cheap to create. So, what if we simply spun up a completely new database for each and every git branch you use in your application? Moreover, what if we could branch and merge SQLite databases easily whenever we needed?

Let's explore how we can leverage the fact that the `database.yml` file is parsed by ERB before being consumed as well as the fact that SQLite affords a number of tools to work with and manipulate database files. With a mix of a Rails `after_initialize` hook and a new bin/ script, we can create a powerful workflow for working with separate databases for each git branch in our project.