Error monitoring
We need not stick to only the Rails provided database-backed adapters for various features that need or would benefit from having a persistent data store. The pattern of using dedicated SQLite databases to back persistent-data-needing components is generally useful. For example, we could connect this with Rails' error reporting interface to build an integrated error monitoring service. This is precisely what the Solid Errors gem does. Install and setup the gem to explore what integrated error monitoring can look like.